requirements.txt
openai~=0.28.0
python_dotenv
requests
beautifulsoup4